Added user group setting editor form
